feat: using generated dtos instead of type assertion
